Move prefetch Blitz query to blitz/next
